refactor(student): 詳細分析を AnalyticsPanel として切り出す
AnalyticsContent が StudentDashboard の内側で定義されていた。内側で
定義したコンポーネントは親が再描画されるたびに別物として作り直される
ため、state が毎回消え、分析データを取り直していた。

- src/components/student/AnalyticsPanel.js へ移動（480行）
- 親から要るのはタブ移動・教材選択・学習開始の3つだけなので props で渡す
- レベルの呼び名を src/config/levels.json に統一。この画面だけ 1〜10 の
  独自対応表を持っていて、実データに1語も存在しない英検1級まで並んで
  いた。使わなくなった getEikenLevel は削除
- 「データエラー」の判定も 10 から MAX_WORD_LEVEL(7) に合わせる

StudentDashboard は 3,189 → 2,727 行。
